refactor(rpc): extract secondary-index operation handlers
Move secondary-index read, write, delete, and update operation handling into secondary_index_operations.cc.
Register secondary_index_operations.cc in the server build.
