+++ server/CMakeLists.txt
@@ -63,6 +63,7 @@ set(SOURCES
     rpc/read_plan_executor.cc
     rpc/row_codec.cc
     rpc/row_codec.hh
+    rpc/secondary_index_operations.cc
     rpc/stateless_operations.cc
     rpc/table_stats.cc
     rpc/transaction_control.cc
